backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
numpy
PyPDF2
python-docx
beautifulsoup4
sentence-transformers
pyttsx3
motor
google-genai
